LayoutTests/http/tests/security/cross-frame-access-put-expected.txt

CONSOLE MESSAGE: line 151: Blocked a frame with origin "http://localhost:8000" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
ALERT: PASS: window.Attr should be 'function Attr() {    [native code]}' and is.
ALERT: PASS: window.CDATASection should be 'function CDATASection() {    [native code]}' and is.
ALERT: PASS: window.CharacterData should be 'function CharacterData() {    [native code]}' and is.
ALERT: PASS: window.Comment should be 'function Comment() {    [native code]}' and is.
ALERT: PASS: window.CSSPrimitiveValue should be 'function CSSPrimitiveValue() {    [native code]}' and is.
ALERT: PASS: window.CSSRule should be 'function CSSRule() {    [native code]}' and is.
ALERT: PASS: window.CSSStyleDeclaration should be 'function CSSStyleDeclaration() {    [native code]}' and is.
ALERT: PASS: window.CSSValue should be 'function CSSValue() {    [native code]}' and is.
ALERT: PASS: window.Document should be 'function Document() {    [native code]}' and is.
ALERT: PASS: window.DocumentFragment should be 'function DocumentFragment() {    [native code]}' and is.
ALERT: PASS: window.DocumentType should be 'function DocumentType() {    [native code]}' and is.
ALERT: PASS: window.DOMException should be 'function DOMException() {    [native code]}' and is.
ALERT: PASS: window.DOMImplementation should be 'function DOMImplementation() {    [native code]}' and is.
ALERT: PASS: window.DOMParser should be 'function DOMParser() {    [native code]}' and is.
ALERT: PASS: window.Element should be 'function Element() {    [native code]}' and is.
ALERT: PASS: window.EvalError should be 'function EvalError() {    [native code]}' and is.
ALERT: PASS: window.Event should be 'function Event() {    [native code]}' and is.
ALERT: PASS: window.HTMLAnchorElement should be 'function HTMLAnchorElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLAppletElement should be 'function HTMLAppletElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLAreaElement should be 'function HTMLAreaElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLBaseElement should be 'function HTMLBaseElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLBodyElement should be 'function HTMLBodyElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLBRElement should be 'function HTMLBRElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLButtonElement should be 'function HTMLButtonElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLCanvasElement should be 'function HTMLCanvasElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLDirectoryElement should be 'function HTMLDirectoryElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLDivElement should be 'function HTMLDivElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLDListElement should be 'function HTMLDListElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLDocument should be 'function HTMLDocument() {    [native code]}' and is.
ALERT: PASS: window.HTMLElement should be 'function HTMLElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLFieldSetElement should be 'function HTMLFieldSetElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLFontElement should be 'function HTMLFontElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLFormElement should be 'function HTMLFormElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLFrameElement should be 'function HTMLFrameElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLFrameSetElement should be 'function HTMLFrameSetElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLHeadElement should be 'function HTMLHeadElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLHeadingElement should be 'function HTMLHeadingElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLHRElement should be 'function HTMLHRElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLHtmlElement should be 'function HTMLHtmlElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLIFrameElement should be 'function HTMLIFrameElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLImageElement should be 'function HTMLImageElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLInputElement should be 'function HTMLInputElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLIsIndexElement should be 'undefined' and is.
ALERT: PASS: window.HTMLLabelElement should be 'function HTMLLabelElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLLegendElement should be 'function HTMLLegendElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLLIElement should be 'function HTMLLIElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLLinkElement should be 'function HTMLLinkElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLMapElement should be 'function HTMLMapElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLMarqueeElement should be 'function HTMLMarqueeElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLMenuElement should be 'function HTMLMenuElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLMetaElement should be 'function HTMLMetaElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLModElement should be 'function HTMLModElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLOListElement should be 'function HTMLOListElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLOptGroupElement should be 'function HTMLOptGroupElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLOptionElement should be 'function HTMLOptionElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLParagraphElement should be 'function HTMLParagraphElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLParamElement should be 'function HTMLParamElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLPreElement should be 'function HTMLPreElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLQuoteElement should be 'function HTMLQuoteElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLScriptElement should be 'function HTMLScriptElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLSelectElement should be 'function HTMLSelectElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLStyleElement should be 'function HTMLStyleElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTableCaptionElement should be 'function HTMLTableCaptionElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTableCellElement should be 'function HTMLTableCellElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTableColElement should be 'function HTMLTableColElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTableElement should be 'function HTMLTableElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTableRowElement should be 'function HTMLTableRowElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTableSectionElement should be 'function HTMLTableSectionElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTextAreaElement should be 'function HTMLTextAreaElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLTitleElement should be 'function HTMLTitleElement() {    [native code]}' and is.
ALERT: PASS: window.HTMLUListElement should be 'function HTMLUListElement() {    [native code]}' and is.
ALERT: PASS: window.MutationEvent should be 'function MutationEvent() {    [native code]}' and is.
ALERT: PASS: window.Node should be 'function Node() {    [native code]}' and is.
ALERT: PASS: window.NodeFilter should be 'function NodeFilter() {    [native code]}' and is.
ALERT: PASS: window.ProcessingInstruction should be 'function ProcessingInstruction() {    [native code]}' and is.
ALERT: PASS: window.Range should be 'function Range() {    [native code]}' and is.
ALERT: PASS: window.RangeError should be 'function RangeError() {    [native code]}' and is.
ALERT: PASS: window.RangeException should be 'undefined' and is.
ALERT: PASS: window.ReferenceError should be 'function ReferenceError() {    [native code]}' and is.
ALERT: PASS: window.SyntaxError should be 'function SyntaxError() {    [native code]}' and is.
ALERT: PASS: window.Text should be 'function Text() {    [native code]}' and is.
ALERT: PASS: window.TypeError should be 'function TypeError() {    [native code]}' and is.
ALERT: PASS: window.URIError should be 'function URIError() {    [native code]}' and is.
ALERT: PASS: window.XMLDocument should be 'function XMLDocument() {    [native code]}' and is.
ALERT: PASS: window.XMLSerializer should be 'function XMLSerializer() {    [native code]}' and is.
ALERT: PASS: window.XPathEvaluator should be 'function XPathEvaluator() {    [native code]}' and is.
ALERT: PASS: window.XPathResult should be 'function XPathResult() {    [native code]}' and is.
ALERT: PASS: window.clientInformation should be '[object Navigator]' and is.
ALERT: PASS: window.closed should be 'false' and is.
ALERT: PASS: window.console should be '[object Console]' and is.
ALERT: PASS: window.crypto should be '[object Crypto]' and is.
ALERT: PASS: window.defaultStatus should be '' and is.
ALERT: PASS: window.defaultstatus should be '' and is.
ALERT: PASS: window.devicePixelRatio should be '1' and is.
ALERT: PASS: window.document should be '[object HTMLDocument]' and is.
ALERT: PASS: window.embeds should be 'undefined' and is.
ALERT: PASS: window.event should be 'undefined' and is.
CONSOLE MESSAGE: line 14: Blocked a frame with origin "http://localhost:8000" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
ALERT: PASS: window.frameElement should be 'null' and is.
ALERT: PASS: window.frames should be '[object Window]' and is.
ALERT: PASS: window.history should be '[object History]' and is.
ALERT: PASS: window.images should be 'undefined' and is.
ALERT: PASS: window.innerHeight should be '150' and is.
ALERT: PASS: window.innerWidth should be '300' and is.
ALERT: PASS: window.length should be '0' and is.
ALERT: PASS: window.locationbar should be '[object BarProp]' and is.
ALERT: PASS: window.menubar should be '[object BarProp]' and is.
ALERT: PASS: window.name should be '' and is.
ALERT: PASS: window.navigator should be '[object Navigator]' and is.
ALERT: PASS: window.offscreenBuffering should be 'true' and is.
ALERT: PASS: window.onabort should be 'null' and is.
ALERT: PASS: window.onbeforeunload should be 'null' and is.
ALERT: PASS: window.onblur should be 'null' and is.
ALERT: PASS: window.onchange should be 'null' and is.
ALERT: PASS: window.onclick should be 'null' and is.
ALERT: PASS: window.ondblclick should be 'null' and is.
ALERT: PASS: window.onerror should be 'null' and is.
ALERT: PASS: window.onfocus should be 'null' and is.
ALERT: PASS: window.onkeydown should be 'null' and is.
ALERT: PASS: window.onkeypress should be 'null' and is.
ALERT: PASS: window.onkeyup should be 'null' and is.
ALERT: PASS: window.onload should be 'null' and is.
ALERT: PASS: window.onmousedown should be 'null' and is.
ALERT: PASS: window.onmousemove should be 'null' and is.
ALERT: PASS: window.onmouseout should be 'null' and is.
ALERT: PASS: window.onmouseover should be 'null' and is.
ALERT: PASS: window.onmouseup should be 'null' and is.
ALERT: PASS: window.onmousewheel should be 'null' and is.
ALERT: PASS: window.onreset should be 'null' and is.
ALERT: PASS: window.onresize should be 'null' and is.
ALERT: PASS: window.onscroll should be 'null' and is.
ALERT: PASS: window.onsearch should be 'null' and is.
ALERT: PASS: window.onselect should be 'null' and is.
ALERT: PASS: window.onsubmit should be 'null' and is.
ALERT: PASS: window.onunload should be 'null' and is.
ALERT: PASS: window.opener should be 'null' and is.
ALERT: PASS: window.outerHeight matched the expected value.
ALERT: PASS: window.outerWidth matched the expected value.
ALERT: PASS: window.pageXOffset should be '0' and is.
ALERT: PASS: window.pageYOffset should be '0' and is.
ALERT: PASS: window.personalbar should be '[object BarProp]' and is.
ALERT: PASS: window.plugins should be 'undefined' and is.
ALERT: PASS: window.screen should be '[object Screen]' and is.
ALERT: PASS: window.screenLeft should be '0' and is.
ALERT: PASS: window.screenTop matched the expected value.
ALERT: PASS: window.screenX should be '0' and is.
ALERT: PASS: window.screenY matched the expected value.
ALERT: PASS: window.scrollbars should be '[object BarProp]' and is.
ALERT: PASS: window.scrollX should be '0' and is.
ALERT: PASS: window.scrollY should be '0' and is.
ALERT: PASS: window.self should be '[object Window]' and is.
ALERT: PASS: window.status should be '' and is.
ALERT: PASS: window.statusbar should be '[object BarProp]' and is.
ALERT: PASS: window.toolbar should be '[object BarProp]' and is.
ALERT: PASS: window.window should be '[object Window]' and is.
ALERT: PASS: window.parent should be parentOld and is.
ALERT: PASS: window.top should be topOld and is.
ALERT: PASS: window.addEventListener should be 'function addEventListener() {    [native code]}' and is.
ALERT: PASS: window.alert should be 'function alert() {    [native code]}' and is.
ALERT: PASS: window.atob should be 'function atob() {    [native code]}' and is.
ALERT: PASS: window.btoa should be 'function btoa() {    [native code]}' and is.
ALERT: PASS: window.captureEvents should be 'function captureEvents() {    [native code]}' and is.
ALERT: PASS: window.clearInterval should be 'function clearInterval() {    [native code]}' and is.
ALERT: PASS: window.clearTimeout should be 'function clearTimeout() {    [native code]}' and is.
ALERT: PASS: window.confirm should be 'function confirm() {    [native code]}' and is.
ALERT: PASS: window.eval should be 'function eval() {    [native code]}' and is.
ALERT: PASS: window.find should be 'function find() {    [native code]}' and is.
ALERT: PASS: window.getComputedStyle should be 'function getComputedStyle() {    [native code]}' and is.
ALERT: PASS: window.getMatchedCSSRules should be 'function getMatchedCSSRules() {    [native code]}' and is.
ALERT: PASS: window.getSelection should be 'function getSelection() {    [native code]}' and is.
ALERT: PASS: window.moveBy should be 'function moveBy() {    [native code]}' and is.
ALERT: PASS: window.moveTo should be 'function moveTo() {    [native code]}' and is.
ALERT: PASS: window.open should be 'function open() {    [native code]}' and is.
ALERT: PASS: window.print should be 'function print() {    [native code]}' and is.
ALERT: PASS: window.prompt should be 'function prompt() {    [native code]}' and is.
ALERT: PASS: window.releaseEvents should be 'function releaseEvents() {    [native code]}' and is.
ALERT: PASS: window.removeEventListener should be 'function removeEventListener() {    [native code]}' and is.
ALERT: PASS: window.resizeBy should be 'function resizeBy() {    [native code]}' and is.
ALERT: PASS: window.resizeTo should be 'function resizeTo() {    [native code]}' and is.
ALERT: PASS: window.scroll should be 'function scroll() {    [native code]}' and is.
ALERT: PASS: window.scrollBy should be 'function scrollBy() {    [native code]}' and is.
ALERT: PASS: window.scrollTo should be 'function scrollTo() {    [native code]}' and is.
ALERT: PASS: window.setInterval should be 'function setInterval() {    [native code]}' and is.
ALERT: PASS: window.setTimeout should be 'function setTimeout() {    [native code]}' and is.
ALERT: PASS: window.showModalDialog matched the expected value.
ALERT: PASS: window.stop should be 'function stop() {    [native code]}' and is.
CONSOLE MESSAGE: line 38: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 39: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 40: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 41: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 42: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 43: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 44: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 45: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 46: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 47: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 48: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 49: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 50: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 51: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 52: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 53: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 54: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 55: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 56: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 57: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 58: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 59: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 60: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 61: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 62: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 63: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 64: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 65: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 66: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 67: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 68: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 69: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 70: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 71: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 72: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 73: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 74: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 75: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 76: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 77: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 78: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 79: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 80: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 81: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 82: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 83: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 84: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 85: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 86: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 87: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 88: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 89: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 90: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 91: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 92: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 93: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 94: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 95: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 96: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 97: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 98: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 99: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 100: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 101: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 102: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 103: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 104: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 105: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 106: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 107: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 108: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 109: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 110: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 111: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 112: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 113: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 114: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 115: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 116: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 117: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 118: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 119: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 120: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 121: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 122: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 131: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 133: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 135: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 136: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 137: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 138: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 139: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 140: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 142: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 143: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 144: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 145: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 146: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 147: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 148: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 149: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 150: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 151: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 152: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 153: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 154: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 155: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 156: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 157: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 158: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 159: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 160: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 161: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 162: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 163: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 164: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 165: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 166: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 167: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 168: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 169: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 170: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 171: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 172: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 173: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 174: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 175: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 176: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 177: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 178: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 179: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 180: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 183: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 184: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 185: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 186: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 187: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 188: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 189: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 190: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 191: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 192: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 193: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 194: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 195: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 196: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 197: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 198: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 201: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 202: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 203: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 204: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 205: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 206: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 207: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 208: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 209: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 210: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 211: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 212: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 213: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 214: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 215: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 216: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 217: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 218: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 219: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 220: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 221: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 222: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 223: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 224: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 225: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 226: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 227: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 228: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 229: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 230: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 231: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 232: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
CONSOLE MESSAGE: line 233: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
This test checks cross-frame access security of window attribute setters (rdar://problem/5326791).


MAIN WINDOW: !!-- Test ended--!!
